Make the ecf module build again
Make the ecf module build again.
Snappy does not come with a pc file anymore,
use the cmake config that comes with it
instead. Don't use SNAPPY_INCLUDE_DIR it
comes with the Snappy::snappy target.
Adapt the code to build with newer qpp that
supports doxygen 0.9.5+.
Due to that this now requires qore 1.12.4
don't test for qore versions just compile
using C++11.